move buffer to /puller
