+++ cdc/puller/buffer_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kv
+package puller
 
 import (
 	"context"
@@ -19,6 +19,7 @@ import (
 	"time"
 
 	"github.com/pingcap/check"
+	"github.com/pingcap/tidb-cdc/cdc/kv"
 	"github.com/pingcap/tidb-cdc/pkg/util"
 )
 
@@ -45,11 +46,11 @@ func (bs *bufferSuite) TestCanAddAndReadEntriesInOrder(c *check.C) {
 		c.Assert(third.KV.Ts, check.Equals, uint64(112))
 	}()
 
-	err := b.AddKVEntry(ctx, &RawKVEntry{Ts: 111})
+	err := b.AddKVEntry(ctx, &kv.RawKVEntry{Ts: 111})
 	c.Assert(err, check.IsNil)
 	err = b.AddResolved(ctx, util.Span{}, 110)
 	c.Assert(err, check.IsNil)
-	err = b.AddKVEntry(ctx, &RawKVEntry{Ts: 112})
+	err = b.AddKVEntry(ctx, &kv.RawKVEntry{Ts: 112})
 	c.Assert(err, check.IsNil)
 
 	wg.Wait()
@@ -63,7 +64,7 @@ func (bs *bufferSuite) TestWaitsCanBeCanceled(c *check.C) {
 	defer cancel()
 	stopped := make(chan struct{})
 	go func() {
-		err := b.AddEntry(timeout, BufferEntry{KV: &RawKVEntry{Ts: 111}})
+		err := b.AddEntry(timeout, BufferEntry{KV: &kv.RawKVEntry{Ts: 111}})
 		c.Assert(err, check.Equals, context.DeadlineExceeded)
 		close(stopped)
 	}()
